+++ src/Exception.php
@@ -2,7 +2,7 @@
 namespace Nikolag\Square;
 
 use Illuminate\Queue\SerializesModels;
-use \Exception as BaseException;
+use Nikolag\Core\Exceptions\Exception as BaseException;
 
 class Exception extends BaseException {
 	use SerializesModels;
@@ -15,7 +15,7 @@ class Exception extends BaseException {
 	 * @param BaseException|null $previous 
 	 * @return void
 	 */
-	public function __construct($message = null, $code = 0, BaseException $previous = null)
+	public function __construct($message = null, $code = 0, Exception $previous = null)
 	{
 		//Parent
 		parent::__construct($message, $code, $previous);