+++ src/contracts/SquareServiceContract.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Nikolag\Square\Contracts;
+
+use Nikolag\Core\Contracts\PaymentServiceContract;
+
+interface SquareServiceContract extends PaymentServiceContract
+{
+    
+}