+++ src/utils/Constants.php
@@ -2,17 +2,9 @@
 
 namespace Nikolag\Square\Utils;
 
-class Constants {
-	//Transaction info
-	const TRANSACTION_NAMESPACE 	= 'Nikolag\Square\Models\Transaction';
-	const TRANSACTION_IDENTIFIER 	= 'id';
-	//Transaction statuses
-	const TRANSACTION_STATUS_OPENED = 'PENDING';
-	const TRANSACTION_STATUS_PASSED = 'PAID';
-	const TRANSACTION_STATUS_FAILED = 'FAILED';
-	//Customer info
-	const CUSTOMER_NAMESPACE 		= 'Nikolag\Square\Models\Customer';
-	const CUSTOMER_IDENTIFIER 		= 'id';
+use Nikolag\Core\Utils\Constants as CoreConstants;
+
+class Constants extends CoreConstants {
 	//Exceptions
 	//INVALID_REQUEST_ERROR
 	const INVALID_REQUEST_ERROR 	= 'INVALID_REQUEST_ERROR';